test: restructure lib tests (#63)
